e2e(renode): consolidate dya2 foundation — robust connect + per-boot execution
Merge the dya2 tab fan-out (tab-a: home/subsystems; tab-b: keymap/macro-combo +
robust connect helper) and harden the shared pieces for the wave-2 fan-out.

- Canonical connectDya2 (tests/dya2/dya2.helpers.ts) now uses the robust
  connect: PREFER dya-studio's page-load auto-reconnect (the WebSerial shim
  always reports the paired port, so the app reconnects with no click) and only
  click "Connect via USB" as a fallback. Clicking during auto-reconnect opened
  the shared serial port twice and desynced RPC framing (every request after
  GetDeviceInfo stalled — the empty Subsystems list). Deleted the separate
  connect-robust.helper.ts; keymap/macro-combo now import from ./dya2.helpers.
  Documented both dya2 facts (auto-reconnect, one Studio connection per boot) +
  the per-boot local recipe at the top of dya2.helpers.ts.

- subsystems.spec.ts: the enumeration worked all along — all dya2 subsystems are
  "supported", so they render inside the collapsed "Already supported by DYA
  Studio" SectionCard (children only mount when expanded). Expand that section
  before counting. Settled EXPECTED_SUBSYSTEM_IDENTIFIERS to a confirmed-present
  curated subset (the DUT advertises 14; kscan-diagnostics is NOT shipped).

- Per-boot execution: dya2's two-machine wired-split DUT serves only ONE Studio
  connection per boot, so each spec must run in its own Renode boot. Added a
  per-DUT `one_connection_per_boot` matrix flag; the e2e job branches on it —
  official keeps its single-boot run, dya2 loops run-local.sh once per spec
  (expanding tests/dya2 to individual *.spec.ts). Documented in firmware/README.

Verified against the real dya2 DUT in Renode, each spec in its own boot:
connect, subsystems, home, keymap, macro-combo all PASS (runtime-combo stays
test.fixme-skipped).
